+++ bin/find-all-logs.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+find wp-content/uploads -iname 'splinter*.html'