+++ inc/File_Logger.php
@@ -64,10 +64,10 @@ class File_Logger{
 	 * Writes the log details to the log file
 	 *
 	 * @param string $title
-	 * @param string $content
+	 * @param string $info
 	 * @param string $log_level
 	 */
-	private static function add( $title = '', $content = '', $log_level =  self::LOG_LEVEL_STANDARD ){
+	private static function add( $title = '', $info = '', $log_level =  self::LOG_LEVEL_STANDARD ){
 		
 		if( !$title ){
 			return;
@@ -75,26 +75,23 @@ class File_Logger{
 		
 		self::prep();
 		
-		if( is_array( $content ) || is_object($content) ){
-			$formatted_content = sprintf('<div><details closed><summary>Content</summary><pre>%s</pre></details></div>',var_export( $content, true ));
+		if( is_array( $info ) || is_object($info) ){
+			$formatted_info = sprintf('<div><details closed><summary>Content</summary><pre>%s</pre></details></div>',var_export( $info, true ));
 		}else{
-			$formatted_content = strval($content);
+			$formatted_info = strval($info);
 		}
 		
 		$stacktrace = sprintf('<div><details closed><summary>%s</summary><pre>%s</pre></details></div>','Stacktrace', var_export(wp_debug_backtrace_summary(null,null,false), true));;
 		
-		$html = '<table class="table m-0">
-		    <tbody>
+		$html = '
 		      <tr class="table-%s">
-		    	<td><span style="display:block;width:170px">%s</span></td>
-		    	<td>%s</td>
-				<td>%s</td>
-				<td>%s</td>
-			</tr>
-		    </tbody>
-		  </table>';
+		    	<td class="timestamp"><div style="width:170px">%s</div></td>
+		    	<td class="message">%s</td>
+				<td class="info" style="max-width:50vw">%s</td>
+				<td class="stacktrace">%s</td>
+			</tr>';
 		
-		$html = sprintf($html, $log_level, date_i18n( 'Y-m-d H:i' ), $title, $formatted_content, $stacktrace );
+		$html = sprintf($html, $log_level, date_i18n( 'Y-m-d H:i' ), $title, $formatted_info, $stacktrace );
 		
 		
 		self::write($html);
@@ -117,60 +114,75 @@ class File_Logger{
 	 * Creates a standard log entry
 	 *
 	 * @param string $title
-	 * @param string $content
+	 * @param string $info
 	 */
-	public static function log( $title = '', $content = '' ){
-		self::add( $title, $content, self::LOG_LEVEL_STANDARD );
+	public static function log( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_STANDARD );
 	}
 	
 	/**
 	 * Creates a standard log entry
 	 *
 	 * @param string $title
-	 * @param string $content
+	 * @param string $info
 	 */
-	public static function info( $title = '', $content = '' ){
-		self::add( $title, $content, self::LOG_LEVEL_INFO );
+	public static function info( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_INFO );
 	}
 	
 	/**
 	 * Creates a standard log entry
 	 *
 	 * @param string $title
-	 * @param string $content
+	 * @param string $info
 	 */
-	public static function error( $title = '', $content = '' ){
-		self::add( $title, $content, self::LOG_LEVEL_ERROR );
+	public static function error( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_ERROR );
 	}
 	
-	
-	public static function debug( $title = '', $content = '' ){
-		self::add( $title, $content, self::LOG_LEVEL_DEBUG );
+	/**
+	 * Creates a debug log entry
+	 *
+	 * @param string $title
+	 * @param string $info
+	 */
+	public static function debug( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_DEBUG );
 	}
 	
-	
-	public static function warn( $title = '', $content = '' ){
-		self::add( $title, $content, self::LOG_LEVEL_WARNING );
+	/**
+	 * Creates a warning log entry
+	 *
+	 * @param string $title
+	 * @param string $info
+	 */
+	public static function warn( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_WARNING );
 	}
 	
 	/**
 	 * Creates a primary log entry
 	 *
 	 * @param string $title
+	 * @param string $info
 	 */
-	public static function primary( $title = '' ){
-		self::add( $title, null, self::LOG_LEVEL_PRIMARY );
+	public static function primary( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_PRIMARY );
 	}
 	
 	/**
 	 * Creates a log entry with a success level
 	 *
 	 * @param string $title
+	 * @param string $info
 	 */
-	public static function success( $title = '' ){
-		self::add( $title, null, self::LOG_LEVEL_SUCCESS );
+	public static function success( $title = '', $info = '' ){
+		self::add( $title, $info, self::LOG_LEVEL_SUCCESS );
 	}
 	
+	/**
+	 * Preps the log file with the necessary headers
+	 */
 	private static function prep() {
 		if ( file_exists( self::get_writable_file() ) ) {
 			return;
@@ -187,14 +199,31 @@ class File_Logger{
 				<th><details closed><summary>$_SERVER</summary><pre style="color:white" class="font-weight-light">%s</pre></details></th>
 			</tr>
 		    </thead>
-		  </table>';
+		    <tbody>';
 			$header .= sprintf($html, var_export($_SERVER, true) );
 		}
 		self::write($header);
 	}
 	
-	public function get_current_log_path(){
+	/**
+	 * Returns the current log file being used
+	 *
+	 * @return false|string
+	 */
+	public static function get_current_log_path(){
 		return self::$current_log_file;
+	}
+	
+	/**
+	 * Logs all super globals to the current log file
+	 */
+	public static function log_all_globals(){
+		self::debug( '$_POST', $_POST );
+		self::debug( '$_GET', $_GET );
+		self::debug( '$_FILES', $_POST );
+		self::debug( '$_SESSION', $_POST );
+		self::debug( '$_SERVER', $_POST );
+		self::debug( '$_REQUEST', $_POST );
 	}
 }
 
@@ -206,6 +235,7 @@ add_action( __NAMESPACE__.'\File_Logger\primary', ['\Splinter\File_Logger','prim
 add_action( __NAMESPACE__.'\File_Logger\log', ['\Splinter\File_Logger','log'], 15, 2 );
 add_action( __NAMESPACE__.'\File_Logger\info', ['\Splinter\File_Logger','info'], 15, 2 );
 add_action( __NAMESPACE__.'\File_Logger\error', ['\Splinter\File_Logger','error'], 15, 2 );
-add_action( __NAMESPACE__.'\File_Logger\warn', ['\Splinter\File_Logger','error'], 15, 2 );
-add_action( __NAMESPACE__.'\File_Logger\debug', ['\Splinter\File_Logger','warn'], 15, 2 );
+add_action( __NAMESPACE__.'\File_Logger\warn', ['\Splinter\File_Logger','warn'], 15, 2 );
+add_action( __NAMESPACE__.'\File_Logger\debug', ['\Splinter\File_Logger','debug'], 15, 2 );
 add_action( __NAMESPACE__.'\File_Logger\success', ['\Splinter\File_Logger','success'], 15, 2 );
+add_action( __NAMESPACE__.'\File_Logger\log_all_globals', ['\Splinter\File_Logger','log_all_globals'], 15, 2 );