+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # 2018-08-22
 * Initial Commit
 * Initialized as composer package
+* Fixed static function ``get_current_log_file()``
+* Added function ``log_all_globals`` and an accompanying action
+* Updated Doc Blocks
+* Fixed table formatting on the log output